library_backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.7.0
boto3==1.29.4
botocore==1.32.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-auth==0.9.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.13.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
msgpack==1.0.7
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.4.0
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.25.0
zope.interface==6.1
